Fix wrong cell dtype

+++ mlcg/neighbor_list/torch_impl.py
@@ -110,7 +110,7 @@ def compute_images(
 ) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
     """TODO: add doc"""
 
-    cell = cell.view((-1, 3, 3)).to(torch.float64)
+    cell = cell.view((-1, 3, 3))
     pbc = pbc.view((-1, 3))
     reciprocal_cell = torch.linalg.inv(cell).transpose(2, 1)
     # print('reciprocal_cell: ', reciprocal_cell.device)
